requirements.txt
django==1.9.6
djangorestframework==3.3.3
djangorestframework-filters==0.8.0
django-debug-toolbar==1.4

jsonfield==1.0.3

markdown==2.6.6
